+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* added snapshot tx
+* added snapshot mode in noninteractive transaction
 
 ## 1.12.0
 * added StaticAuthentication