+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* fix keepInCache param in YdbQuery
 * added Yson type
 * add logger as Ydb config
 * added snapshot mode in noninteractive transaction