+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## 1.13.0
 * added transaction mode for retryTransaction
 * fix keepInCache param in YdbQuery
 * added Yson type